Tratativas de Error

+++ src/tasks/tasks.controller.ts
@@ -6,7 +6,10 @@ import {
   Param,
   Post,
   Put,
+  HttpException,
+  HttpStatus,
 } from '@nestjs/common';
+
 import { Task } from './shared/task';
 import { TaskService } from './shared/task.service';
 
@@ -20,7 +23,13 @@ export class TasksController {
   }
 
   @Get(':id')
-  async getById(@Param('id') id: string): Promise<Task> {
+  async getById(@Param('id') id: string): Promise<Task | number> {
+    const countTasks = await this.taskService.getById(id);
+
+    if (countTasks == 0) {
+      throw new HttpException('ID not found', HttpStatus.NOT_FOUND);
+    }
+
     return this.taskService.getById(id);
   }
 
@@ -30,12 +39,25 @@ export class TasksController {
   }
 
   @Put(':id')
-  async update(@Param('id') id: string, @Body() task: Task): Promise<Task> {
+  async update(
+    @Param('id') id: string,
+    @Body() task: Task,
+  ): Promise<Task | number> {
+    const countTasks = await this.taskService.getById(id);
+
+    if (countTasks == 0) {
+      throw new HttpException('ID not found', HttpStatus.NOT_FOUND);
+    }
     return this.taskService.update(id, task);
   }
 
   @Delete(':id')
   async delete(@Param('id') id: string): Promise<void> {
+    const countTasks = await this.taskService.getById(id);
+
+    if (countTasks == 0) {
+      throw new HttpException('ID not found', HttpStatus.NOT_FOUND);
+    }
     this.taskService.delete(id);
   }
 }
